fix: #579 google_mapが上手く動かないことがあったため、stimulusを使って実装するように変更
