+++ app/views/home/index.html.erb
@@ -1,8 +1,8 @@
 <div id="home_index" class="bg-white">
   <div class="mx-auto max-w-1920">
-    <div class="p-6 space-y-6">
+    <div data-controller="google-map" class="p-6 space-y-6">
       <h2>撮影スポットマップ</h2>
-      <div id="map" class="w-full h-full aspect-[3/1] min-h-[50vh] rounded-lg" data-posts=<%= @posts_json %>></div>
+      <div data-google-map-target="map" class="w-full h-full aspect-[3/1] min-h-[50vh] rounded-lg" data-posts=<%= @posts_json %>></div>
     </div>
     <div class="p-6 space-y-6">
       <div class="flex items-center justify-between">
