+++ app/views/posts/_form.html.erb
@@ -7,17 +7,17 @@
       <%= image_tag select_image_preview(post.image, "large"), data: { image_preview_target: "image" }, class:"h-auto max-w-full max-h-[500px]" %>
     </div>
   </div>
-  <div class="space-y-2">
+  <div data-controller="google-map" class="space-y-2">
     <%= f.label :location %>
     <div class="flex gap-4">
-      <%= f.text_field :location, class: "text-field" %>
-      <button type="button" id="mapSearch" class="map-button">地図で検索</button>
+      <%= f.text_field :location, class: "text-field", data: { google_map_target: "location" } %>
+      <button data-action="google-map#codeAddress" type="button" class="map-button">地図で検索</button>
     </div>
-    <div id="map" class="h-72 md:h-[600px]"></div>
+    <div data-google-map-target="map" class="h-72 md:h-[600px]"></div>
     <%= f.label :latitude, class: "sr-only" %>
-    <%= f.hidden_field :latitude %>
+    <%= f.hidden_field :latitude, data: { google_map_target: "latitude" } %>
     <%= f.label :longitude, class: "sr-only" %>
-    <%= f.hidden_field :longitude %>
+    <%= f.hidden_field :longitude, data: { google_map_target: "longitude" } %>
   </div>
   <div class="space-y-2">
     <%= f.label :title %>
